refactor(roles): extract PermissionsService from RolesService (verification vs management)
Moves verifyActionsForUserAndResource (RBAC hot path: instance fast path,
per-resource-type community resolution, private-channel and DM membership
checks) into a dedicated PermissionsService depending only on DatabaseService.
RbacGuard now injects PermissionsService; the method is deleted from
RolesService (single external call site, no delegate kept).

roles.service.ts 1523 -> 1343 lines; verification describe blocks moved from
roles.service.spec.ts into permissions.service.spec.ts (585 lines).

Verified: 122 suites / 2064 tests green, type-check + lint clean, backend
boots with 0 compile errors.
