+++ backend/src/roles/roles.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { RolesService } from './roles.service';
+import { PermissionsService } from './permissions.service';
 import { DatabaseModule } from '@/database/database.module';
 import { RolesController } from './roles.controller';
 
 @Module({
   imports: [DatabaseModule],
-  providers: [RolesService],
-  exports: [RolesService],
+  providers: [RolesService, PermissionsService],
+  exports: [RolesService, PermissionsService],
   controllers: [RolesController],
 })
 export class RolesModule {}
